feat: implemented the global exceptional handling with traceId logging

+++ src/main/java/com/talentprobe/assessment/service/UserService.java
@@ -5,6 +5,8 @@ import com.talentprobe.assessment.dto.UserResponseDto;
 import com.talentprobe.assessment.entity.User;
 import com.talentprobe.assessment.enums.Role;
 import com.talentprobe.assessment.enums.Status;
+import com.talentprobe.assessment.exception.DuplicateResourceException;
+import com.talentprobe.assessment.exception.ResourceNotFoundException;
 import com.talentprobe.assessment.mapper.UserMapper;
 import com.talentprobe.assessment.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -13,6 +15,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.UUID;
+
 @Service
 @RequiredArgsConstructor
 public class UserService {
@@ -22,7 +25,7 @@ public class UserService {
 
     public UserResponseDto createCandidate(UserDto dto) {
         if (userRepository.existsByEmail(dto.getEmail())) {
-            throw new RuntimeException("Email already registered");
+            throw new DuplicateResourceException("Email already registered");
         }
         User user = userMapper.toEntity(dto);
         user.setRole(Role.CANDIDATE);
@@ -40,20 +43,22 @@ public class UserService {
 
     public UserResponseDto getUserById(UUID id) {
         User user = userRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("User not found"));
+                .orElseThrow(() -> new ResourceNotFoundException("User not found"));
         return userMapper.toDto(user);
     }
 
     public UserResponseDto updateUser(UUID id, UserDto dto) {
         User user = userRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("User not found"));
+                .orElseThrow(() -> new ResourceNotFoundException("User not found"));
         user.setName(dto.getName());
         user.setPhoneNumber(dto.getPhoneNumber());
         return userMapper.toDto(userRepository.save(user));
     }
 
     public void deleteUser(UUID id) {
+        if (!userRepository.existsById(id)) {
+            throw new ResourceNotFoundException("User not found");
+        }
         userRepository.deleteById(id);
     }
-
 }
